flagships/contract_redline_with_provenance/pyproject.toml
[project]
name = "ks-cookbook-contract-redline"
version = "0.1.0"
description = "Legal flagship — produce a Word redline memo where every proposed edit is bound to both the offending clause chunk and the governing playbook rule chunk."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[mcp]>=0.0.14",
    "pydantic>=2.6",
    "python-docx>=1.1",
]

[project.scripts]
ks-cookbook-contract-redline = "contract_redline.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contract_redline"]
